+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/CallTarget.java
@@ -40,24 +40,68 @@
  */
 package com.oracle.truffle.api;
 
+import com.oracle.truffle.api.nodes.DirectCallNode;
+import com.oracle.truffle.api.nodes.EncapsulatingNodeReference;
+import com.oracle.truffle.api.nodes.IndirectCallNode;
+import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.RootNode;
 
 /**
- * Represents the target of a call. Do not subclass {@link CallTarget} directly but implement
- * {@link RootNode} instead and retrieve its {@link CallTarget target} via
- * {@link RootNode#getCallTarget()}.
+ * Represents the target of a call. Call targets are created automatically from a {@link RootNode}
+ * by calling {@link RootNode#getCallTarget()}.
+ * <p>
+ * A call target allows the runtime to employ a compilation heuristic to trigger partial evaluation
+ * of the underlying {@link RootNode}, typically in the background. Additionally, calling a call
+ * target builds a guest language level {@link TruffleStackTrace stack trace} which can be inspected
+ * using {@link TruffleRuntime#iterateFrames(com.oracle.truffle.api.frame.FrameInstanceVisitor)} or
+ * {@link TruffleStackTrace#getStackTrace(Throwable)}.
+ * <p>
+ * Do not subclass {@link CallTarget} directly, as this interface is likely to become sealed in the
+ * future.
  *
  * @see RootNode
+ * @see DirectCallNode
+ * @see IndirectCallNode
  * @since 0.8 or earlier
  */
 public interface CallTarget {
 
     /**
-     * Calls this target as a root method.
+     * Calls the encapsulated root node with the given arguments and returns the result.
+     * <p>
+     * By calling this method, the call location is looked up using
+     * {@link EncapsulatingNodeReference}. Use {@link #call(Node, Object...)} if the call location
+     * is already known.
+     * <p>
+     * Calling this method in partially evaluated code will allow it to get inlined if the receiver
+     * (this) is a {@link CompilerDirectives#isPartialEvaluationConstant(Object) pe-constant}. Call
+     * site {@link RootNode#isCloningAllowed() cloning} is only supported if a
+     * {@link DirectCallNode} is used instead.
      *
-     * @param arguments passed arguments as an object array
-     * @return the return result of the call
+     * @param arguments The arguments passed to the call, as an object array.
+     * @return The result of the call.
+     * @see #call(Node, Object...)
      * @since 0.8 or earlier
      */
     Object call(Object... arguments);
+
+    /**
+     * Calls the encapsulated root node with an explicit call location and arguments, and returns
+     * the result.
+     * <p>
+     * This method should be preferred over {@link #call(Object...)} if the current location is
+     * known, as it avoids looking up the current location from a thread-local.
+     * <p>
+     *
+     * @param location A {@link Node} that identifies the location of this call. The location may be
+     *            <code>null</code> if no location is available.
+     * @param arguments The arguments passed to the call, as an object array.
+     * @return The result of the call.
+     * @see #call(Object...)
+     * @since 24.1
+     */
+    default Object call(Node location, Object... arguments) {
+        throw CompilerDirectives.shouldNotReachHere("callDirect not supported for this runtime");
+    }
+
 }