+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleStackTrace.java
@@ -56,6 +56,7 @@ import com.oracle.truffle.api.frame.FrameInstanceVisitor;
 import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.nodes.ControlFlowException;
 import com.oracle.truffle.api.nodes.Node;
+import com.oracle.truffle.api.nodes.RootNode;
 
 /**
  * Represents a guest language stack trace.
@@ -76,7 +77,7 @@ import com.oracle.truffle.api.nodes.Node;
  * <p>
  * See {@link #getStackTrace(Throwable)} to retrieve the guest language stack trace from a
  * {@link Throwable}.
- * 
+ *
  * @since 19.0
  */
 @SuppressWarnings("serial")
@@ -213,6 +214,9 @@ public final class TruffleStackTrace extends Exception {
         if (throwable instanceof ControlFlowException) {
             return EMPTY;
         }
+        if (throwable instanceof PolyglotException) {
+            return EMPTY;
+        }
         LazyStackTrace lazy = getOrCreateLazyStackTrace(throwable);
         if (lazy.stackTrace != null) {
             // stack trace already exists
@@ -241,7 +245,8 @@ public final class TruffleStackTrace extends Exception {
         for (ListIterator<TracebackElement> iterator = elements.listIterator(elements.size()); iterator.hasPrevious();) {
             TracebackElement element = iterator.previous();
             if (element.root != null) {
-                frames.add(new TruffleStackTraceElement(topCallSite, element.root, element.frame));
+                int bytecodeIndex = LanguageAccessor.NODES.findBytecodeIndex(element.root.getRootNode(), topCallSite, element.frame);
+                frames.add(new TruffleStackTraceElement(topCallSite, element.root, element.frame, bytecodeIndex));
                 topCallSite = null;
             }
             if (element.callNode != null) {
@@ -251,7 +256,7 @@ public final class TruffleStackTrace extends Exception {
         int lazyFrames = frames.size();
 
         // attach the remaining stack trace elements
-        addStackFrames(stackFrameLimit, lazyFrames, topCallSite, frames);
+        addFramesByStackWalking(stackFrameLimit, topCallSite, frames);
 
         TruffleStackTrace fullStackTrace = new TruffleStackTrace(frames, lazyFrames);
         // capture host stack trace for guest language exceptions;
@@ -316,49 +321,43 @@ public final class TruffleStackTrace extends Exception {
         }
     }
 
-    static void addStackFrameInfo(Node callNode, RootCallTarget root, Throwable t, Frame currentFrame) {
+    static void addStackFrameInfo(Node callNode, RootCallTarget target, Throwable t, Frame currentFrame) {
         if (t instanceof ControlFlowException) {
             // control flow exceptions should never have to get a stack trace.
             return;
         }
-        if (t instanceof PolyglotException) {
-            // Normally, polyglot exceptions should not even end up here, with the exception of
-            // those thrown by Value call targets. In any case, we do not want to attach a cause.
-            return;
-        }
-
-        boolean isTProfiled = CompilerDirectives.isPartialEvaluationConstant(t.getClass());
         MaterializedFrame frame = null;
-        if (currentFrame != null && root.getRootNode().isCaptureFramesForTrace()) {
+        if (currentFrame != null && LanguageAccessor.NODES.isCaptureFramesForTrace(target.getRootNode(), callNode)) {
             frame = currentFrame.materialize();
         }
-        callInnerAddStackFrameInfo(isTProfiled, callNode, root, t, frame);
+        callInnerAddStackFrameInfo(callNode, target, t, frame);
     }
 
-    private static void callInnerAddStackFrameInfo(boolean isTProfiled, Node callNode, RootCallTarget root, Throwable t, MaterializedFrame currentFrame) {
-        if (isTProfiled) {
-            innerAddStackFrameInfo(callNode, root, t, currentFrame);
+    private static void callInnerAddStackFrameInfo(Node callNode, RootCallTarget root, Throwable t, MaterializedFrame currentFrame) {
+        boolean isException = LanguageAccessor.EXCEPTIONS.isException(t);
+        if (CompilerDirectives.inCompiledCode() && CompilerDirectives.isPartialEvaluationConstant(isException) && isException) {
+            innerAddStackFrame(callNode, root, t, currentFrame);
         } else {
-            innerAddStackFrameInfoBoundary(callNode, root, t, currentFrame);
+            innerAddStackFrameSlow(callNode, root, t, currentFrame);
         }
     }
 
     @TruffleBoundary
-    private static void innerAddStackFrameInfoBoundary(Node callNode, RootCallTarget root, Throwable t, MaterializedFrame currentFrame) {
-        innerAddStackFrameInfo(callNode, root, t, currentFrame);
-    }
-
-    private static void innerAddStackFrameInfo(Node callNode, RootCallTarget root, Throwable t, MaterializedFrame currentFrame) {
-        if (!LanguageAccessor.EXCEPTIONS.isException(t)) {
+    private static void innerAddStackFrameSlow(Node callNode, RootCallTarget root, Throwable t, MaterializedFrame currentFrame) {
+        if (LanguageAccessor.EXCEPTIONS.isException(t)) {
             /*
              * Capture as much information as possible for internal errors. This branch should not
              * be reached by host exceptions as they should have already been wrapped in a
              * HostException in the guest-to-host call root node.
              */
+            innerAddStackFrame(callNode, root, t, currentFrame);
+        } else {
             fillIn(t);
-            return;
         }
+    }
 
+    private static void innerAddStackFrame(Node callNode, RootCallTarget root, Throwable t, MaterializedFrame currentFrame) {
+        assert LanguageAccessor.EXCEPTIONS.isException(t);
         int stackTraceElementLimit = LanguageAccessor.EXCEPTIONS.getStackTraceElementLimit(t);
         LazyStackTrace lazy = (LazyStackTrace) LanguageAccessor.EXCEPTIONS.getLazyStackTrace(t);
         if (lazy == null) {
@@ -419,7 +418,8 @@ public final class TruffleStackTrace extends Exception {
         }
     }
 
-    private static void addStackFrames(int stackFrameLimit, int lazyFrames, final Node topCallSite, List<TruffleStackTraceElement> frames) {
+    private static void addFramesByStackWalking(int stackFrameLimit, final Node topCallSite, List<TruffleStackTraceElement> frames) {
+        int lazyFrames = frames.size();
         if (stackFrameLimit >= 0 && lazyFrames >= stackFrameLimit) {
             // early exit: avoid costly iterateFrames call if enough frames have been recorded
             // lazily
@@ -435,22 +435,29 @@ public final class TruffleStackTrace extends Exception {
                     // no more frames to create
                     return frameInstance;
                 }
-                Node location = frameInstance.getCallNode();
-                RootCallTarget target = (RootCallTarget) frameInstance.getCallTarget();
+                Node callNode;
                 if (first) {
-                    location = topCallSite;
+                    callNode = topCallSite;
                     first = false;
+                } else {
+                    callNode = frameInstance.getCallNode();
                 }
-                boolean captureFrames = target != null && target.getRootNode().isCaptureFramesForTrace();
-                Frame frame = captureFrames ? frameInstance.getFrame(FrameAccess.READ_ONLY) : null;
-                frames.add(new TruffleStackTraceElement(location, target, frame));
-                first = false;
+
+                RootCallTarget target = ((RootCallTarget) frameInstance.getCallTarget());
+                RootNode root = target.getRootNode();
+                Frame frame = captureFrame(frameInstance, root, callNode);
+                int bytecodeIndex = LanguageAccessor.NODES.findBytecodeIndex(root, callNode, frame);
+                frames.add(new TruffleStackTraceElement(callNode, target, frame, bytecodeIndex));
                 if (target != null && LanguageAccessor.ACCESSOR.nodeSupport().countsTowardsStackTraceLimit(target.getRootNode())) {
                     stackFrameIndex++;
                 }
                 return null;
             }
         });
+    }
+
+    private static Frame captureFrame(FrameInstance frame, RootNode rootNode, Node callNode) {
+        return LanguageAccessor.NODES.isCaptureFramesForTrace(rootNode, callNode) ? frame.getFrame(FrameAccess.READ_ONLY) : null;
     }
 
 }