+++ truffle/src/com.oracle.truffle.runtime/src/com/oracle/truffle/runtime/OptimizedRuntimeAccessor.java
@@ -52,6 +52,7 @@ final class OptimizedRuntimeAccessor extends Accessor {
     static final LanguageSupport LANGUAGE = ACCESSOR.languageSupport();
     static final EngineSupport ENGINE = ACCESSOR.engineSupport();
     static final InteropSupport INTEROP = ACCESSOR.interopSupport();
+    static final ExceptionSupport EXCEPTION = ACCESSOR.exceptionSupport();
     static final FrameSupport FRAME = ACCESSOR.framesSupport();
 
     private OptimizedRuntimeAccessor() {